# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MineVault Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthResult {
  success: boolean;
  userId: number;
  username: string;
}

export interface World {
  id: number;
  name: string;
  /** @nullable */
  ownerId: number | null;
  blockData: string[][];
}

export type GameActionInputActionType = typeof GameActionInputActionType[keyof typeof GameActionInputActionType];


export const GameActionInputActionType = {
  break: 'break',
  place: 'place',
} as const;

export interface GameActionInput {
  actionType: GameActionInputActionType;
  worldName: string;
  x: number;
  y: number;
  /** @nullable */
  activeTool?: string | null;
  /** @nullable */
  placeBlock?: string | null;
}

export interface GameActionResult {
  success: boolean;
  /** @nullable */
  gemsGained?: number | null;
  /** @nullable */
  pointsAwarded?: number | null;
  /** @nullable */
  dropItem?: string | null;
  /** @nullable */
  currentActions?: number | null;
  /** @nullable */
  fatigueLevel?: string | null;
  /** @nullable */
  wizardChallenge?: boolean | null;
}

export interface Miner {
  userId: number;
  level: number;
  unlocked?: boolean;
  currentBalance: number;
  temperature: number;
  isRunning: boolean;
  solarPanels: number;
  generators: number;
  fuel: number;
  lastMaintenanceAt: string;
  lastTickAt: string;
  ratePerSecond?: number;
}

export type MaintainInputType = typeof MaintainInputType[keyof typeof MaintainInputType];


export const MaintainInputType = {
  flush_cooling: 'flush_cooling',
  thermal_paste: 'thermal_paste',
} as const;

export interface MaintainInput {
  type: MaintainInputType;
}

export interface Wallet {
  userId: number;
  gems: number;
  worldLocks: number;
  diamondLocks: number;
  realBalance: number;
  windowPoints: number;
  actionCount: number;
}

export interface InventoryItem {
  itemId: string;
  quantity: number;
  displayName?: string;
  category?: string;
}

export interface StoreItem {
  itemId: string;
  displayName: string;
  gemCost: number;
  /** @nullable */
  realCost?: number | null;
  category: string;
  description: string;
}

export interface BuyItemInput {
  itemId: string;
  quantity: number;
}

export interface PurchaseResult {
  success: boolean;
  message: string;
  /** @nullable */
  newGemBalance?: number | null;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  gems: number;
  windowPoints: number;
  minerLevel: number;
  realBalance?: number;
}

export interface PoolStatus {
  currentPool: number;
  nextPayoutIn: number;
  houseCut: number;
  userCut: number;
  totalWindowPoints?: number;
}

export type MonetizationTaskInputType = typeof MonetizationTaskInputType[keyof typeof MonetizationTaskInputType];


export const MonetizationTaskInputType = {
  drill_boost: 'drill_boost',
  cool_down: 'cool_down',
} as const;

export interface MonetizationTaskInput {
  type: MonetizationTaskInputType;
}

export interface MonetizationTaskResult {
  success: boolean;
  token: string;
  expiresAt: string;
  adUrl?: string;
}

export interface VerifyTaskInput {
  token: string;
}

export interface VerifyTaskResult {
  success: boolean;
  reward: string;
  message?: string;
}