# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MineVault Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Registration result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Unauthorized
...
  /game/action:
    post:
      operationId: gameAction
      tags: [game]
      summary: Perform a game action (break block, place block)
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameActionInput"
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameActionResult"
        "400":
          description: Invalid action
  /miner:
    get:
      operationId: getMiner
      tags: [miner]
      summary: Get current miner state
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Miner state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Miner"
  /miner/tick:
    post:
      operationId: minerTick
      tags: [miner]
      summary: Server-side tick to update miner balance and temperature
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated miner state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Miner"
  /miner/upgrade:
    post:
      operationId: upgradeMiner
      tags: [miner]
      summary: Upgrade miner level
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upgraded miner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Miner"
        "400":
          description: Not enough funds
  /miner/maintain:
    post:
      operationId: maintainMiner
      tags: [miner]
      summary: Apply maintenance (flush cooling / thermal paste)
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintainInput"
      responses:
        "200":
          description: Maintained miner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Miner"
...
components:
  schemas:
...
    AuthInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
    AuthResult:
      type: object
      required: [success, userId, username]
      properties:
        success:
          type: boolean
        userId:
          type: integer
        username:
          type: string
...
    GameActionInput:
      type: object
      required: [actionType, worldName, x, y]
      properties:
        actionType:
          type: string
          enum: [break, place]
        worldName:
          type: string
        x:
          type: integer
        y:
          type: integer
        activeTool:
          type: ["string", "null"]
        placeBlock:
          type: ["string", "null"]
    GameActionResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        gemsGained:
          type: ["integer", "null"]
        pointsAwarded:
          type: ["number", "null"]
        dropItem:
          type: ["string", "null"]
        currentActions:
          type: ["integer", "null"]
        fatigueLevel:
          type: ["string", "null"]
        wizardChallenge:
          type: ["boolean", "null"]
    Miner:
      type: object
      required: [userId, level, currentBalance, temperature, isRunning, solarPanels, generators, fuel, lastMaintenanceAt, lastTickAt]
      properties:
        userId:
          type: integer
        level:
          type: integer
        unlocked:
          type: boolean
        currentBalance:
          type: number
        temperature:
          type: number
        isRunning:
          type: boolean
        solarPanels:
          type: integer
        generators:
          type: integer
        fuel:
          type: integer
        lastMaintenanceAt:
          type: string
        lastTickAt:
          type: string
        ratePerSecond:
          type: number
    MaintainInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [flush_cooling, thermal_paste]
...