# lib/api-zod/src/generated/types/maintainInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MineVault Game API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintainInputType } from './maintainInputType';

export interface MaintainInput {
  type: MaintainInputType;
}