# lib/api-zod/src/generated/types/miner.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MineVault Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Miner {
  userId: number;
  level: number;
  unlocked?: boolean;
  currentBalance: number;
  temperature: number;
  isRunning: boolean;
  solarPanels: number;
  generators: number;
  fuel: number;
  lastMaintenanceAt: string;
  lastTickAt: string;
  ratePerSecond?: number;
}